+++ Breathe/app/src/androidTest/java/com/ybeltagy/breathe/SimpleIueReadWriteTest.java
@@ -7,7 +7,7 @@ import androidx.room.Room;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
-import com.ybeltagy.breathe.persistence.BreatheDao;
+import com.ybeltagy.breathe.data.BreatheDao;
 import com.ybeltagy.breathe.persistence.BreatheRoomDatabase;
 import com.ybeltagy.breathe.data.InhalerUsageEvent;
